Проверка формы и отправка ее
Валидация введенных дааных и отправка теперь (формируется массив)

+++ js/check_cart.js
@@ -5,7 +5,10 @@ function checkCart() {
     //проверяю наличие корзины в localStorage
     if ( localStorage.getItem('cart') != null ) {
         cart = JSON.parse(localStorage.getItem('cart'));
-    } else { cart = []; }
+    } else {
+        cart = [];
+        document.getElementById('ndsfnsdfjsklf').style.display = 'none';
+    }
 }
 
 function showCart() {
@@ -16,21 +19,32 @@ function showCart() {
     if ( rexdiv != undefined ) {
         // если есть элемент корзины, тогда прорисовываем ее
         rexdiv.innerHTML = out;
+        var ids = [];
+        var count = [];
+        var i = 0;
         console.log('===== корзина начало ======== ');
         for (var z in cart) {
-            console.log('отправка запроса на сервер с параметрами id:',z,' Количество:',cart[z]);
-            $.ajax({
-                url: '../ajax_cart.php', // на этот скрипт отправка
-                type: 'post', // метод post
-                response:'text',
-                data: {id: z, kolvo: cart[z]}, // параметры отправки
-                success: function (data) { // выполнение отправки
-                    console.log('заполенение переменной out для прорисовки',data);
-                    out += data;
-                    rexdiv.innerHTML = out;
-                }
-            });
+            if (cart[z] > 0 ) {
+                ids.push(z);
+                count.push(cart[z]);
+                console.log('отправка запроса на сервер с параметрами id:',z,' Количество:',cart[z]);
+                i++;
+
+            }
         }
+        //console.log('Ключи :',ids);
+        //console.log('Количество :',count);
+        $.ajax({
+            url: '../ajax_cart.php', // на этот скрипт отправка
+            type: 'post', // метод post
+            response:'text',
+            data: {id: ids, kolvo: count}, // параметры отправки
+            success: function (data) { // выполнение отправки
+                console.log('заполенение переменной out для прорисовки',data);
+                out += data;
+                rexdiv.innerHTML = out;
+            }
+        });
     }
 }
 
@@ -45,6 +59,7 @@ function showMiniCart() {
     }
     if (all == 0) {
         localStorage.removeItem('cart');
+        document.getElementById('ndsfnsdfjsklf').style.display = 'none';
     }
     cartdiv.innerHTML = 'Корзина: ('+all+')';
 }
@@ -55,26 +70,48 @@ function addCheck() {
     var check_phone = document.getElementById("checkphone").value;
     var check_adress = document.getElementById("checkadress").value;
     var check_komment = document.getElementById("checkkomment").value;
+    var error = 0;
 
-    for (var z in cart) {
-        $.ajax({
-            url: '../ajax_check.php', // на этот скрипт отправка
-            type: 'post', // метод post
-            response: 'text',
-            data: {id: z, kolvo: cart[z], name: checkname, phone: check_phone, adress: check_adress, komment: check_komment}, // параметры отправки
-            success: function (data) { // выполнение отправки
-                console.log('Ответ от сервера', data);
-            }
-        });
-    }
-    localStorage.clear();
-    checkCart();
-    showCart();
-    showMiniCart();
-    alert('Спасибо Вам ' + checkname + ', Ваш заказ принят, ожидайте звонка нашего менеджера и смс на Ваш телефон.');
-    document.getElementById("checkname").value = '';
-    document.getElementById("checkphone").value= '';
-    document.getElementById("checkadress").value= '';
-    document.getElementById("checkkomment").value= '';
+     if (checkname.length < 3) {
+         error = 1;
+         document.getElementById('labelname').innerHTML = "<font color='red'>Имя слишком короткое</font>";
+     } else {
+         document.getElementById('labelname').innerHTML = "<font color='green'>Имя полученно</font>";
+     }
+     if (check_phone.length !== 16) {
+         error = 1;
+         document.getElementById('labelphone').innerHTML = "<font color='red'>Номер не полный</font>";
+     } else {
+         document.getElementById('labelphone').innerHTML = "<font color='green'>Телефон получен</font>";
+     }
+     if ( check_adress < 3 ) {
+         error = 1;
+         document.getElementById('labeladress').innerHTML = "<font color='red'>Адрес слишком короткий</font>";
+     } else {
+         document.getElementById('labeladress').innerHTML = "<font color='green'>Адресс получен</font>";
+     }
+     if (error == 0) {
+         for (var z in cart) {
+             $.ajax({
+                 url: '../ajax_check.php', // на этот скрипт отправка
+                 type: 'post', // метод post
+                 response: 'text',
+                 data: {id: z, kolvo: cart[z], name: checkname, phone: check_phone, adress: check_adress, komment: check_komment}, // параметры отправки
+                 success: function (data) { // выполнение отправки
+                     console.log('Ответ от сервера', data);
+                 }
+             });
+         }
+         localStorage.clear();
+         checkCart();
+         showCart();
+         showMiniCart();
+         alert('Спасибо Вам ' + checkname + ', Ваш заказ принят, ожидайте звонка нашего менеджера и смс на Ваш телефон.');
+         //alert(data);
+         document.getElementById("checkname").value = '';
+         document.getElementById("checkphone").value= '';
+         document.getElementById("checkadress").value= '';
+         document.getElementById("checkkomment").value= '';
+     }
 
 }
